requirements.txt
argcomplete==2.0.0
attrs==22.2.0
Babel==2.10.3
blinker==1.5
boto3==1.26.27
botocore==1.29.27
Brlapi==0.8.4; sys_platform == 'linux'
certifi==2022.9.24
chardet==5.1.0
charset-normalizer>=2,<3
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
cupshelpers==1.0; sys_platform == 'linux'
dbus-python==1.3.2; sys_platform == 'linux'
distro==1.8.0
distro-info==1.5+deb12u1; sys_platform == 'linux'
ghp-import==2.1.0
httplib2==0.20.4
idna==3.3
iniconfig==1.1.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kafka-python==2.0.2
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
livereload==2.6.3
louis==3.24.0; sys_platform == 'linux'
lunr==0.6.2
Mako==1.2.4
Markdown>=3.2.1,<3.4
MarkupSafe==2.1.2
mergedeep==1.3.4
mkdocs==1.4.2
more-itertools==8.10.0
nltk==3.8
numpy==1.24.2
oauthlib==3.2.2
packaging==23.0
pipx==1.1.0
pluggy==1.0.0
psutil==5.9.4
py==1.11.0
pycairo==1.20.1; sys_platform == "linux"
pycups==2.0.1; sys_platform == "linux"
pycurl==7.45.2; sys_platform == 'linux'
Pygments==2.14.0
PyGObject==3.42.2; sys_platform == "linux"
pyinotify==0.9.6; sys_platform == "linux"
PyJWT==2.6.0
pyparsing==3.0.9
PySimpleSOAP==1.16.2
pysmbc==1.0.23; sys_platform == "linux"
pytest==7.2.1
python-apt==2.6.0; sys_platform == "linux"
python-dateutil==2.8.2
python-debian==0.1.49
python-debianbts==4.0.1
pytz==2022.7.1
pyxdg==0.28
PyYAML==6.0
pyyaml_env_tag==0.1
regex==2022.10.31
reportbug==12.0.0; sys_platform == "linux"
requests==2.28.1
s3transfer==0.6.0
simplejson==3.18.3
six==1.16.0
tornado==6.2
tqdm==4.64.1
ufw==0.36.2; sys_platform == "linux"
urllib3==1.26.12
userpath==1.8.0
wadllib==1.3.6
watchdog==2.2.1
xdg==5
